// app/src/generated-api/models/CreateDriverRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Travel Manager - API de Gestão de recursos de transporte rodoviário
 * Especificação modular da API de back-end. Todos os schemas e paths estão em ficheiros separados (specs/). Use um agregador OpenAPI (como o ReDoc ou Swagger UI) para visualizar a documentação completa.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ferdxavier@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Dados necessários para criar ou atualizar um motorista (sem o ID gerado).
 * @export
 * @interface CreateDriverRequest
 */
export interface CreateDriverRequest {
    /**
     * Nome do motorista.
     * @type {string}
     * @memberof CreateDriverRequest
     */
    name: string;
    /**
     * Número do CPF (apenas 11 dígitos).
     * @type {string}
     * @memberof CreateDriverRequest
     */
    cpf: string;
    /**
     * Número do RG ou Carteira de Identidade.
     * @type {string}
     * @memberof CreateDriverRequest
     */
    rg?: string;
    /**
     * Data de Nascimento (YYYY-MM-DD). Requer validação de idade (+18) no código da API.
     * @type {Date}
     * @memberof CreateDriverRequest
     */
    dateBirth: Date;
    /**
     * Número da Carteira Nacional de Habilitação (CNH).
     * @type {string}
     * @memberof CreateDriverRequest
     */
    cnhNumber: string;
    /**
     * Data de validade da CNH. Requer validação para ser no futuro no código da API.
     * @type {Date}
     * @memberof CreateDriverRequest
     */
    validityCnh: Date;
    /**
     * Data de validade do exame toxicológico. Requer validação para ser no futuro no código da API.
     * @type {Date}
     * @memberof CreateDriverRequest
     */
    validityToxicological: Date;
}

/**
 * Check if a given object implements the CreateDriverRequest interface.
 */
export function instanceOfCreateDriverRequest(value: object): value is CreateDriverRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('cpf' in value) || value['cpf'] === undefined) return false;
    if (!('dateBirth' in value) || value['dateBirth'] === undefined) return false;
    if (!('cnhNumber' in value) || value['cnhNumber'] === undefined) return false;
    if (!('validityCnh' in value) || value['validityCnh'] === undefined) return false;
    if (!('validityToxicological' in value) || value['validityToxicological'] === undefined) return false;
    return true;
}

export function CreateDriverRequestFromJSON(json: any): CreateDriverRequest {
    return CreateDriverRequestFromJSONTyped(json, false);
}

export function CreateDriverRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateDriverRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'cpf': json['cpf'],
        'rg': json['rg'] == null ? undefined : json['rg'],
        'dateBirth': (new Date(json['dateBirth'])),
        'cnhNumber': json['cnhNumber'],
        'validityCnh': (new Date(json['validityCnh'])),
        'validityToxicological': (new Date(json['validityToxicological'])),
    };
}

export function CreateDriverRequestToJSON(json: any): CreateDriverRequest {
    return CreateDriverRequestToJSONTyped(json, false);
}

export function CreateDriverRequestToJSONTyped(value?: CreateDriverRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'cpf': value['cpf'],
        'rg': value['rg'],
        'dateBirth': value['dateBirth'].toISOString().substring(0,10),
        'cnhNumber': value['cnhNumber'],
        'validityCnh': value['validityCnh'].toISOString().substring(0,10),
        'validityToxicological': value['validityToxicological'].toISOString().substring(0,10),
    };
}